+++ lib/domain/record/components/button/cancel_button.dart
@@ -4,33 +4,34 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:pilll/analytics.dart';
 import 'package:pilll/components/atoms/buttons.dart';
 import 'package:pilll/domain/record/record_page_store.dart';
-import 'package:pilll/domain/record/util/take.dart';
 import 'package:pilll/entity/pill_sheet.dart';
-import 'package:pilll/util/datetime/day.dart';
 
-class TakenButton extends HookWidget {
-  final BuildContext context;
+class CancelButton extends HookWidget {
   final PillSheet pillSheet;
 
-  const TakenButton({
-    Key? key,
-    required this.context,
-    required this.pillSheet,
-  }) : super(key: key);
-  @override
+  CancelButton(this.pillSheet);
   Widget build(BuildContext context) {
     final store = useProvider(recordPageStoreProvider);
-    return PrimaryButton(
-      text: "飲んだ",
-      onPressed: () async {
-        if (pillSheet.todayPillNumber == 1)
-          analytics.logEvent(name: "user_taken_first_day_pill");
-        analytics.logEvent(name: "taken_button_pressed", parameters: {
+    return TertiaryButton(
+      text: "飲んでない",
+      onPressed: () {
+        analytics.logEvent(name: "cancel_taken_button_pressed", parameters: {
           "last_taken_pill_number": pillSheet.lastTakenPillNumber,
           "today_pill_number": pillSheet.todayPillNumber,
         });
-        await take(context, pillSheet, now(), store);
+        _cancelTake(pillSheet, store);
       },
     );
+  }
+
+  void _cancelTake(PillSheet pillSheet, RecordPageStore store) {
+    if (pillSheet.todayPillNumber != pillSheet.lastTakenPillNumber) {
+      return;
+    }
+    final lastTakenDate = pillSheet.lastTakenDate;
+    if (lastTakenDate == null) {
+      return;
+    }
+    store.take(lastTakenDate.subtract(Duration(days: 1)));
   }
 }
